RootSignatureBindable rework (#3)
- Now we make pipeline hold rootindexes instead of bindables, this allows for clear design and structure.
- Added GraphicsBufferSuballocator that allows for allocating data at runtime
- Reworked Index and Vertex buffers to use new GraphicsBufferSuballocator to allow updating meshes at runtime
- Added revision number to bindables, this allows us to act on it in every renderjob individually
- Added Cleanup to ResourceList, this allows us to clear elements that are not used
- Fixed sphere bounding box not updating after chaning mesh

+++ Src/Graphics/Core/Pipeline.h
@@ -8,14 +8,14 @@ class Camera;
 struct DestinationBufferRegionCopyData
 {
 	GraphicsBuffer* buffer;
-	unsigned int byteOffset;
+	size_t byteOffset;
 };
 
 struct SourceBufferRegionCopyData
 {
 	GraphicsBuffer* buffer;
-	unsigned int byteOffset;
-	unsigned int byteSizeToCopy;
+	size_t byteOffset;
+	size_t byteSizeToCopy;
 };
 
 class Pipeline
